Carve out src/graph/ package (closes #8 — Tier B)
Creates src/graph/ as the canonical home for the shared graph
foundation. Two files moved with git mv:

- src/new-abm/models/network.py → src/graph/network.py
  (RoadNode, RoadEdge, RoadNetwork dataclasses)
- src/new-abm/data_generation/graph_vocab.py → src/graph/vocab.py

src/graph/vocab.py's internal import updated to the canonical path
(from graph.network import ...). The new test file uses canonical
imports (from graph.vocab, from graph.network), proving the new
layout is self-supporting.

Backward-compat shims left at the old paths. Each is a 10-line file
that adds src/ to sys.path (if not already there) and re-exports the
public API:

- src/new-abm/models/network.py — re-exports from graph.network
- src/new-abm/data_generation/graph_vocab.py — re-exports from graph.vocab

This means the 22 callers of `from models.network import ...` and 3
callers of `from data_generation.graph_vocab import ...` keep working
unchanged. Shims are temporary; they get deleted in a follow-up after
consumers migrate to the canonical `from graph.X import ...` form.

Scope deferred from the original #8 RFC:
- Renaming src/new-abm → src/simulation (Tier C). Touches
  notebooks/datathon_submission.ipynb (locked submission), README.md,
  scripts/build_colab_notebook.py, visualization/abm_animation/
  export_trajectories.py. Will land in a follow-up with explicit
  byte-identity verification of File_1/2/3.csv outputs.
- Topology and geo_ops module extractions (their own follow-ons).
- Splitting spanish_network.py into graph/simulation halves (its own
  follow-on).

Memory updated: project_structure, task_board, decisions_log.

Tests: 89 passed.
